+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.3.0
 * Magazines are now only retained once the weapon's mastering is maxed out
 * Added a setting to turn the mastering requirement off
+* Added a setting to disable magazine retention
 
 ## v1.2.0
 * Compatibility update for SPT 4.1